requirements.txt
streamlit
python-dotenv
sendgrid
pypdf
openai
agents
